requirements.txt
flask
gunicorn
openai==2.8.1
PyPDF2
python-docx
numpy
pandas
tqdm
sentence-transformers
pinecone-client
tiktoken
langchain-openai
langchain-pinecone
deep-translator
openai
python-dotenv
fuzzywuzzy
python-Levenshtein